Snow and rain improvements
Now snow color is melted with terrain texture, so it's possible to see the morphology of the terrain even when the snow is saturated.
Increased rain particles and changed rain color.
Increased initial height of the snow (now the snowfall is more realistic).
Changed "wet effect" on wet shader.
Increased gravity